requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.3
PyPDF2==2.11.1
python-docx==0.8.11
